Set RESET vector to bootstart
Also sets the jump address to 0x0000 + offset of image_hdr.

This fixes and issue where the MCU would attempt to begin executing the
image_hdr and would thus not function as expected. Now, on reset the MCU
begins executing the bootloading, and when it sees that MCUSR is not 0,
meaning there was a reset due to power-on, hardware, brown-out, or
watchdog, it will immediately just jump to the beginning of the
application.